requirement.txt
streamlit==1.24.0
Pillow==9.5.0
numpy==1.24.3
opencv-python-headless==4.7.0.72
pytesseract==0.3.10
groq==0.3.0